servidorOcr/pyproject.toml
[project]
name = "servidorocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "paddleocr>=3.2.0",
    "paddlepaddle>=3.2.0",
    "pillow>=11.3.0",
    "transformers>=4.57.6",
]
